terraform/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
PyPDF2==3.0.1
pdfplumber==0.11.7
clickhouse-connect==0.8.18
python-dotenv==1.1.1
reportlab==4.4.3
Pillow==11.3.0
PyMuPDF==1.25.1
requests==2.32.4
pytest==8.3.4
pytest-asyncio==0.25.3
httpx==0.28.1
psutil==6.1.1
celery==5.4.0
redis==5.2.1
aioredis==2.0.1
prometheus-client==0.21.1